refactor(checker): split lazy flow env mirror helpers
